requirements.txt
addict==2.4.0
asgi-correlation-id==3.2.0
  starlette==0.20.4
    anyio==3.6.2
      idna==3.2
      sniffio==1.2.0
    typing_extensions==4.4.0
asgiref==3.4.1
astroid==2.8.6
  lazy-object-proxy==1.7.1
  setuptools==60.7.1
  typing_extensions==4.4.0
  wrapt==1.13.3
asyncpg==0.24.0
bcrypt==4.0.1
cfgv==3.3.1
contextvars==2.4
  immutables==0.16
distlib==0.3.6
email-validator==1.1.3
  dnspython==2.1.0
  idna==3.2
factory-boy==3.2.0
  Faker==13.15.1
    python-dateutil==2.8.2
      six==1.16.0
fakeredis==1.9.3
  redis==4.3.4
    async-timeout==4.0.2
    Deprecated==1.2.13
      wrapt==1.13.3
    packaging==21.3
      pyparsing==3.0.7
  sortedcontainers==2.4.0
fastapi-utils==0.2.1
  fastapi==0.86.0
    pydantic==1.10.2
      typing_extensions==4.4.0
    starlette==0.20.4
      anyio==3.6.2
        idna==3.2
        sniffio==1.2.0
      typing_extensions==4.4.0
  pydantic==1.10.2
    typing_extensions==4.4.0
  SQLAlchemy==1.4.25
    greenlet==1.1.2
fastapi-versioning==0.10.0
  fastapi==0.86.0
    pydantic==1.10.2
      typing_extensions==4.4.0
    starlette==0.20.4
      anyio==3.6.2
        idna==3.2
        sniffio==1.2.0
      typing_extensions==4.4.0
  starlette==0.20.4
    anyio==3.6.2
      idna==3.2
      sniffio==1.2.0
    typing_extensions==4.4.0
filelock==3.7.0
freezegun==1.2.1
  python-dateutil==2.8.2
    six==1.16.0
gunicorn==20.1.0
  setuptools==60.7.1
httptools==0.2.0
identify==2.5.1
importlib-metadata==4.8.1
  zipp==3.6.0
mccabe==0.6.1
newrelic==7.2.4.171
nodeenv==1.6.0
orjson==3.7.7
passlib==1.7.4
phonenumbers==8.12.36
pipdeptree==2.2.1
  pip==22.3.1
pipreqs==0.4.10
  docopt==0.6.2
  yarg==0.1.9
    requests==2.26.0
      certifi==2021.10.8
      charset-normalizer==2.0.6
      idna==3.2
      urllib3==1.26.7
platformdirs==2.5.2
pydantic-factories==1.4.1
  Faker==13.15.1
    python-dateutil==2.8.2
      six==1.16.0
  pydantic==1.10.2
    typing_extensions==4.4.0
  typing_extensions==4.4.0
  xeger==0.3.5
pyflakes==2.4.0
pytest-asyncio==0.20.1
  pytest==7.0.1
    attrs==21.4.0
    iniconfig==1.1.1
    packaging==21.3
      pyparsing==3.0.7
    pluggy==1.0.0
    py==1.11.0
    tomli==2.0.1
pytest-cov==3.0.0
  coverage==6.3.2
  pytest==7.0.1
    attrs==21.4.0
    iniconfig==1.1.1
    packaging==21.3
      pyparsing==3.0.7
    pluggy==1.0.0
    py==1.11.0
    tomli==2.0.1
pytest-httpx==0.21.0
  httpx==0.23.0
    certifi==2021.10.8
    httpcore==0.15.0
      anyio==3.6.2
        idna==3.2
        sniffio==1.2.0
      certifi==2021.10.8
      h11==0.12.0
      sniffio==1.2.0
    rfc3986==1.5.0
    sniffio==1.2.0
  pytest==7.0.1
    attrs==21.4.0
    iniconfig==1.1.1
    packaging==21.3
      pyparsing==3.0.7
    pluggy==1.0.0
    py==1.11.0
    tomli==2.0.1
pytest-mock==3.7.0
  pytest==7.0.1
    attrs==21.4.0
    iniconfig==1.1.1
    packaging==21.3
      pyparsing==3.0.7
    pluggy==1.0.0
    py==1.11.0
    tomli==2.0.1
python-dotenv==0.19.0
python-json-logger==2.0.2
python-multipart==0.0.5
  six==1.16.0
pytz==2021.3
remove-print-statements==0.5.0
  click==8.1.3
  libcst==0.4.7
    PyYAML==5.4.1
    typing_extensions==4.4.0
    typing-inspect==0.8.0
      mypy-extensions==0.4.3
      typing_extensions==4.4.0
requests-mock==1.9.3
  requests==2.26.0
    certifi==2021.10.8
    charset-normalizer==2.0.6
    idna==3.2
    urllib3==1.26.7
  six==1.16.0
sentry-sdk==1.5.0
  certifi==2021.10.8
  urllib3==1.26.7
snowflake-id==0.0.2
sqlparse==0.4.2
toml==0.10.2
typer==0.7.0
  click==8.1.3
uvicorn==0.19.0
  click==8.1.3
  h11==0.12.0
uvloop==0.16.0
watchgod==0.7
websockets==10.0
wheel==0.37.1
